csv_agent/requirements.txt
langchain==0.3.10
langchain-community==0.3.10
langchain-experimental==0.3.3
anthropic==0.40.0
ctransformers==0.2.27
langchain-anthropic==0.3.0
pandas==2.2.3
python-dotenv==1.0.1
tiktoken==0.8.0
tabulate==0.9.0
openllm==0.6.15

